pyproject.toml
[project]
name = "mcbe-ai-agent"
version = "2.0.0"
description = "MCBE WebSocket AI Agent - 现代化异步 AI Agent 框架"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "rice-awa" }]
keywords = ["minecraft", "bedrock", "websocket", "AI", "ai-agent", "pydantic"]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pydantic-ai>=1.0.0",
    "websockets>=12.0",
    "httpx>=0.27",
    "PyJWT>=2.8",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.9",
]
anthropic = ["anthropic>=0.25"]
ollama = ["ollama>=0.2"]

[project.scripts]
mcbe-AI = "mcbe_ai_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcbe_ai_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
